+++ src/rust/src/conversion_r_to_s.rs
@@ -212,7 +212,7 @@ fn recursive_robjname2series_tree(x: &Robj, name: &str) -> pl::PolarsResult<Seri
             //todo this could probably in fewer allocations
             let dt = pl::DataType::Datetime(pl::TimeUnit::Milliseconds, tz);
             Ok(SeriesTree::Series(
-                (s.cast(&pl::DataType::Int64)? * 1_000i64).cast(&dt)?,
+                ((s * 1000f64).cast(&pl::DataType::Int64)?).cast(&dt)?,
             ))
         }
         Ok(SeriesTree::Series(s)) if x.inherits("Date") => {